+++ src/CSMapLayer.js
@@ -1,0 +1,5 @@
+/* Provides a layer handler to give consistent data import etc */
+
+'use strict';
+var L = global.L || require('leaflet');
+//import 'leaflet';