+++ src/CSMapOSMLayer.js
@@ -1,0 +1,16 @@
+import 'leaflet';
+import * as turf from "@turf/turf";
+import * as turfHelper from "@turf/helpers";
+import { isUndefined } from 'util';
+import './CSMapLayer';
+
+/* Provides a map layer based on data from OSM */
+export var CSMapOSMLayer = CSMapLayer.extend({
+
+ 
+  }); 
+  
+ 
+  
+  
+